Activity id as Long

+++ Strava API v3/src/javastrava/api/v3/model/StravaActivity.java
@@ -21,11 +21,11 @@ import javastrava.cache.StravaCacheable;
  * @author Dan Shannon
  *
  */
-public class StravaActivity implements StravaCacheable<Integer>{
+public class StravaActivity implements StravaCacheable<Long>{
 	/**
 	 * Strava's unique identifier for the activity
 	 */
-	private Integer id;
+	private Long id;
 	/**
 	 * State of the resource (summary, detailed, etc.)
 	 */
@@ -920,7 +920,7 @@ public class StravaActivity implements StravaCacheable<Integer>{
 	 * @return the id
 	 */
 	@Override
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 	/**
@@ -1346,7 +1346,7 @@ public class StravaActivity implements StravaCacheable<Integer>{
 	/**
 	 * @param id the id to set
 	 */
-	public void setId(final Integer id) {
+	public void setId(final Long id) {
 		this.id = id;
 	}
 	/**
